+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+* Added new test for table and column name limits [#133](https://github.com/cloudquery/cq-provider-sdk/pull/133)
 
 ## [v0.5.5] - 2021-12-15